Promote network security intercept resources to GA. (#13884)
[upstream:b01dee97afa28a160c1716aaa69ffc71f6aa5963]
