+++ .changelog/13884.txt
@@ -1,0 +1,3 @@
+```release-note:note
+networksecurity: promoted `google_network_security_intercept_deployment`, `google_network_security_intercept_deployment_group`, `google_network_security_intercept_endpoint_group_association`, `google_network_security_intercept_endpoint_group` resources from Beta to GA
+```
