+++ google-beta/services/networksecurity/resource_network_security_intercept_deployment_generated_test.go
@@ -1,17 +1,36 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file in
+//     .github/CONTRIBUTING.md.
+//
+// ----------------------------------------------------------------------------
+
 package networksecurity_test
 
 import (
+	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/plancheck"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func TestAccNetworkSecurityInterceptDeployment_update(t *testing.T) {
+func TestAccNetworkSecurityInterceptDeployment_networkSecurityInterceptDeploymentBasicExample(t *testing.T) {
 	t.Parallel()
 
 	context := map[string]interface{}{
@@ -20,45 +39,30 @@ func TestAccNetworkSecurityInterceptDeployment_update(t *testing.T) {
 
 	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
+		CheckDestroy:             testAccCheckNetworkSecurityInterceptDeploymentDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccNetworkSecurityInterceptDeployment_basic(context),
+				Config: testAccNetworkSecurityInterceptDeployment_networkSecurityInterceptDeploymentBasicExample(context),
 			},
 			{
 				ResourceName:            "google_network_security_intercept_deployment.default",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"labels", "terraform_labels"},
-			},
-			{
-				Config: testAccNetworkSecurityInterceptDeployment_update(context),
-				ConfigPlanChecks: resource.ConfigPlanChecks{
-					PreApply: []plancheck.PlanCheck{
-						plancheck.ExpectResourceAction("google_network_security_intercept_deployment.default", plancheck.ResourceActionUpdate),
-					},
-				},
-			},
-			{
-				ResourceName:            "google_network_security_intercept_deployment.default",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"update_time", "labels", "terraform_labels"},
+				ImportStateVerifyIgnore: []string{"intercept_deployment_id", "labels", "location", "terraform_labels"},
 			},
 		},
 	})
 }
 
-func testAccNetworkSecurityInterceptDeployment_basic(context map[string]interface{}) string {
+func testAccNetworkSecurityInterceptDeployment_networkSecurityInterceptDeploymentBasicExample(context map[string]interface{}) string {
 	return acctest.Nprintf(`
 resource "google_compute_network" "network" {
-  provider                = google-beta
   name                    = "tf-test-example-network%{random_suffix}"
   auto_create_subnetworks = false
 }
 
 resource "google_compute_subnetwork" "subnetwork" {
-  provider      = google-beta
   name          = "tf-test-example-subnet%{random_suffix}"
   region        = "us-central1"
   ip_cidr_range = "10.1.0.0/16"
@@ -66,16 +70,14 @@ resource "google_compute_subnetwork" "subnetwork" {
 }
 
 resource "google_compute_region_health_check" "health_check" {
-  provider = google-beta
-  name     = "tf-test-example-hc%{random_suffix}"
-  region   = "us-central1"
+  name   = "tf-test-example-hc%{random_suffix}"
+  region = "us-central1"
   http_health_check {
     port = 80
   }
 }
 
 resource "google_compute_region_backend_service" "backend_service" {
-  provider              = google-beta
   name                  = "tf-test-example-bs%{random_suffix}"
   region                = "us-central1"
   health_checks         = [google_compute_region_health_check.health_check.id]
@@ -84,31 +86,28 @@ resource "google_compute_region_backend_service" "backend_service" {
 }
 
 resource "google_compute_forwarding_rule" "forwarding_rule" {
-  provider               = google-beta
-  name                   = "tf-test-example-fwr%{random_suffix}"
-  region                 = "us-central1"
-  network                = google_compute_network.network.name
-  subnetwork             = google_compute_subnetwork.subnetwork.name
-  backend_service        = google_compute_region_backend_service.backend_service.id
-  load_balancing_scheme  = "INTERNAL"
-  ports                  = [6081]
-  ip_protocol            = "UDP"
+  name                  = "tf-test-example-fwr%{random_suffix}"
+  region                = "us-central1"
+  network               = google_compute_network.network.name
+  subnetwork            = google_compute_subnetwork.subnetwork.name
+  backend_service       = google_compute_region_backend_service.backend_service.id
+  load_balancing_scheme = "INTERNAL"
+  ports                 = [6081]
+  ip_protocol           = "UDP"
 }
 
 resource "google_network_security_intercept_deployment_group" "deployment_group" {
-  provider                      = google-beta
   intercept_deployment_group_id = "tf-test-example-dg%{random_suffix}"
   location                      = "global"
   network                       = google_compute_network.network.id
 }
 
 resource "google_network_security_intercept_deployment" "default" {
-  provider                   = google-beta
   intercept_deployment_id    = "tf-test-example-deployment%{random_suffix}"
   location                   = "us-central1-a"
   forwarding_rule            = google_compute_forwarding_rule.forwarding_rule.id
   intercept_deployment_group = google_network_security_intercept_deployment_group.deployment_group.id
-  description                = "initial description"
+  description                = "some description"
   labels = {
     foo = "bar"
   }
@@ -116,69 +115,41 @@ resource "google_network_security_intercept_deployment" "default" {
 `, context)
 }
 
-func testAccNetworkSecurityInterceptDeployment_update(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-resource "google_compute_network" "network" {
-  provider                = google-beta
-  name                    = "tf-test-example-network%{random_suffix}"
-  auto_create_subnetworks = false
-}
+func testAccCheckNetworkSecurityInterceptDeploymentDestroyProducer(t *testing.T) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		for name, rs := range s.RootModule().Resources {
+			if rs.Type != "google_network_security_intercept_deployment" {
+				continue
+			}
+			if strings.HasPrefix(name, "data.") {
+				continue
+			}
 
-resource "google_compute_subnetwork" "subnetwork" {
-  provider      = google-beta
-  name          = "tf-test-example-subnet%{random_suffix}"
-  region        = "us-central1"
-  ip_cidr_range = "10.1.0.0/16"
-  network       = google_compute_network.network.name
-}
+			config := acctest.GoogleProviderConfig(t)
 
-resource "google_compute_region_health_check" "health_check" {
-  provider = google-beta
-  name     = "tf-test-example-hc%{random_suffix}"
-  region   = "us-central1"
-  http_health_check {
-    port = 80
-  }
-}
+			url, err := tpgresource.ReplaceVarsForTest(config, rs, "{{NetworkSecurityBasePath}}projects/{{project}}/locations/{{location}}/interceptDeployments/{{intercept_deployment_id}}")
+			if err != nil {
+				return err
+			}
 
-resource "google_compute_region_backend_service" "backend_service" {
-  provider              = google-beta
-  name                  = "tf-test-example-bs%{random_suffix}"
-  region                = "us-central1"
-  health_checks         = [google_compute_region_health_check.health_check.id]
-  protocol              = "UDP"
-  load_balancing_scheme = "INTERNAL"
-}
+			billingProject := ""
 
-resource "google_compute_forwarding_rule" "forwarding_rule" {
-  provider               = google-beta
-  name                   = "tf-test-example-fwr%{random_suffix}"
-  region                 = "us-central1"
-  network                = google_compute_network.network.name
-  subnetwork             = google_compute_subnetwork.subnetwork.name
-  backend_service        = google_compute_region_backend_service.backend_service.id
-  load_balancing_scheme  = "INTERNAL"
-  ports                  = [6081]
-  ip_protocol            = "UDP"
-}
+			if config.BillingProject != "" {
+				billingProject = config.BillingProject
+			}
 
-resource "google_network_security_intercept_deployment_group" "deployment_group" {
-  provider                      = google-beta
-  intercept_deployment_group_id = "tf-test-example-dg%{random_suffix}"
-  location                      = "global"
-  network                       = google_compute_network.network.id
-}
+			_, err = transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
+				Config:    config,
+				Method:    "GET",
+				Project:   billingProject,
+				RawURL:    url,
+				UserAgent: config.UserAgent,
+			})
+			if err == nil {
+				return fmt.Errorf("NetworkSecurityInterceptDeployment still exists at %s", url)
+			}
+		}
 
-resource "google_network_security_intercept_deployment" "default" {
-  provider                   = google-beta
-  intercept_deployment_id    = "tf-test-example-deployment%{random_suffix}"
-  location                   = "us-central1-a"
-  forwarding_rule            = google_compute_forwarding_rule.forwarding_rule.id
-  intercept_deployment_group = google_network_security_intercept_deployment_group.deployment_group.id
-  description                = "updated description"
-  labels = {
-    foo = "goo"
-  }
-}
-`, context)
+		return nil
+	}
 }
