+++ google-beta/services/networksecurity/resource_network_security_intercept_endpoint_group_association_generated_test.go
@@ -1,17 +1,36 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file in
+//     .github/CONTRIBUTING.md.
+//
+// ----------------------------------------------------------------------------
+
 package networksecurity_test
 
 import (
+	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/plancheck"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func TestAccNetworkSecurityInterceptEndpointGroupAssociation_update(t *testing.T) {
+func TestAccNetworkSecurityInterceptEndpointGroupAssociation_networkSecurityInterceptEndpointGroupAssociationBasicExample(t *testing.T) {
 	t.Parallel()
 
 	context := map[string]interface{}{
@@ -20,65 +39,47 @@ func TestAccNetworkSecurityInterceptEndpointGroupAssociation_update(t *testing.T
 
 	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
+		CheckDestroy:             testAccCheckNetworkSecurityInterceptEndpointGroupAssociationDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccNetworkSecurityInterceptEndpointGroupAssociation_basic(context),
+				Config: testAccNetworkSecurityInterceptEndpointGroupAssociation_networkSecurityInterceptEndpointGroupAssociationBasicExample(context),
 			},
 			{
 				ResourceName:            "google_network_security_intercept_endpoint_group_association.default",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"labels", "terraform_labels"},
-			},
-			{
-				Config: testAccNetworkSecurityInterceptEndpointGroupAssociation_update(context),
-				ConfigPlanChecks: resource.ConfigPlanChecks{
-					PreApply: []plancheck.PlanCheck{
-						plancheck.ExpectResourceAction("google_network_security_intercept_endpoint_group_association.default", plancheck.ResourceActionUpdate),
-					},
-				},
-			},
-			{
-				ResourceName:            "google_network_security_intercept_endpoint_group_association.default",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"update_time", "labels", "terraform_labels"},
+				ImportStateVerifyIgnore: []string{"intercept_endpoint_group_association_id", "labels", "location", "terraform_labels"},
 			},
 		},
 	})
 }
 
-func testAccNetworkSecurityInterceptEndpointGroupAssociation_basic(context map[string]interface{}) string {
+func testAccNetworkSecurityInterceptEndpointGroupAssociation_networkSecurityInterceptEndpointGroupAssociationBasicExample(context map[string]interface{}) string {
 	return acctest.Nprintf(`
 resource "google_compute_network" "producer_network" {
-  provider                = google-beta
   name                    = "tf-test-example-prod-network%{random_suffix}"
   auto_create_subnetworks = false
 }
 
 resource "google_compute_network" "consumer_network" {
-  provider                = google-beta
   name                    = "tf-test-example-cons-network%{random_suffix}"
   auto_create_subnetworks = false
 }
 
 resource "google_network_security_intercept_deployment_group" "deployment_group" {
-  provider                      = google-beta
   intercept_deployment_group_id = "tf-test-example-dg%{random_suffix}"
   location                      = "global"
   network                       = google_compute_network.producer_network.id
 }
 
 resource "google_network_security_intercept_endpoint_group" "endpoint_group" {
-  provider                      = google-beta
-  intercept_endpoint_group_id   = "tf-test-example-eg%{random_suffix}"
-  location                      = "global"
-  intercept_deployment_group    = google_network_security_intercept_deployment_group.deployment_group.id
+  intercept_endpoint_group_id = "tf-test-example-eg%{random_suffix}"
+  location                    = "global"
+  intercept_deployment_group  = google_network_security_intercept_deployment_group.deployment_group.id
 }
 
 resource "google_network_security_intercept_endpoint_group_association" "default" {
-  provider                                = google-beta
   intercept_endpoint_group_association_id = "tf-test-example-ega%{random_suffix}"
   location                                = "global"
   network                                 = google_compute_network.consumer_network.id
@@ -90,43 +91,41 @@ resource "google_network_security_intercept_endpoint_group_association" "default
 `, context)
 }
 
-func testAccNetworkSecurityInterceptEndpointGroupAssociation_update(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-resource "google_compute_network" "producer_network" {
-  provider                = google-beta
-  name                    = "tf-test-example-prod-network%{random_suffix}"
-  auto_create_subnetworks = false
-}
+func testAccCheckNetworkSecurityInterceptEndpointGroupAssociationDestroyProducer(t *testing.T) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		for name, rs := range s.RootModule().Resources {
+			if rs.Type != "google_network_security_intercept_endpoint_group_association" {
+				continue
+			}
+			if strings.HasPrefix(name, "data.") {
+				continue
+			}
 
-resource "google_compute_network" "consumer_network" {
-  provider                = google-beta
-  name                    = "tf-test-example-cons-network%{random_suffix}"
-  auto_create_subnetworks = false
-}
+			config := acctest.GoogleProviderConfig(t)
 
-resource "google_network_security_intercept_deployment_group" "deployment_group" {
-  provider                      = google-beta
-  intercept_deployment_group_id = "tf-test-example-dg%{random_suffix}"
-  location                      = "global"
-  network                       = google_compute_network.producer_network.id
-}
+			url, err := tpgresource.ReplaceVarsForTest(config, rs, "{{NetworkSecurityBasePath}}projects/{{project}}/locations/{{location}}/interceptEndpointGroupAssociations/{{intercept_endpoint_group_association_id}}")
+			if err != nil {
+				return err
+			}
 
-resource "google_network_security_intercept_endpoint_group" "endpoint_group" {
-  provider                      = google-beta
-  intercept_endpoint_group_id   = "tf-test-example-eg%{random_suffix}"
-  location                      = "global"
-  intercept_deployment_group    = google_network_security_intercept_deployment_group.deployment_group.id
-}
+			billingProject := ""
 
-resource "google_network_security_intercept_endpoint_group_association" "default" {
-  provider                                = google-beta
-  intercept_endpoint_group_association_id = "tf-test-example-ega%{random_suffix}"
-  location                                = "global"
-  network                                 = google_compute_network.consumer_network.id
-  intercept_endpoint_group                = google_network_security_intercept_endpoint_group.endpoint_group.id
-  labels = {
-    foo = "goo"
-  }
-}
-`, context)
+			if config.BillingProject != "" {
+				billingProject = config.BillingProject
+			}
+
+			_, err = transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
+				Config:    config,
+				Method:    "GET",
+				Project:   billingProject,
+				RawURL:    url,
+				UserAgent: config.UserAgent,
+			})
+			if err == nil {
+				return fmt.Errorf("NetworkSecurityInterceptEndpointGroupAssociation still exists at %s", url)
+			}
+		}
+
+		return nil
+	}
 }
