+++ website/docs/r/network_security_intercept_deployment_group.html.markdown
@@ -26,8 +26,6 @@ A deployment group aggregates many zonal intercept backends (deployments)
 into a single global intercept service. Consumers can connect this service
 using an endpoint group.
 
-~> **Warning:** This resource is in beta, and should be used with the terraform-provider-google-beta provider.
-See [Provider Versions](https://terraform.io/docs/providers/google/guides/provider_versions.html) for more details on beta resources.
 
 
 <div class = "oics-button" style="float: right; margin: 0 0 -15px">
@@ -40,13 +38,11 @@ See [Provider Versions](https://terraform.io/docs/providers/google/guides/provid
 
 ```hcl
 resource "google_compute_network" "network" {
-  provider                = google-beta
   name                    = "example-network"
   auto_create_subnetworks = false
 }
 
 resource "google_network_security_intercept_deployment_group" "default" {
-  provider                      = google-beta
   intercept_deployment_group_id = "example-dg"
   location                      = "global"
   network                       = google_compute_network.network.id
